pc-5 refacto search tests into differents functionnals files
