+++ testcafe/05_04_search_by_keyword_and_filters.js
@@ -1,0 +1,22 @@
+// ------------------------ BACK TO SEARCH PAGE
+// Après une recherche si je clique sur back, je reviens à la recherche précédante
+// si je veux revenir à ma page de début de recherche je clique sur la flêche < back et cela réinitialise tous les filtres
+
+// Tu peux cumuler recherche textuelle + filtrage des résultats, et tu peux appliquer un filtre à toute la base d’offres puis faire une recherche textuelle dans la base filtrée
+
+// ------------------------ NAVIGATION PAR TYPE (LIRE, ETC)
+
+// Voir les fonctionnalitées sur https://github.com/betagouv/pass-culture-browser/issues/664
+
+// pour tout rénitialiser, il faut cliquer sur la croix dans le form et sur réninitaliser aussi donc
+
+// Quand on clique sur le bouton chercher, si le filtre est ouvert, il se referme mais la recherche reste active
+
+// ------------------------ BACK TO SEARCH PAGE
+// Après une recherche si je clique sur back, je reviens à la recherche précédante
+// si je veux revenir à ma page de début de recherche je clique sur la flêche < back et cela réinitialise tous les filtres
+
+//  reinitialisation des filters : checkbox unchecked and location.pathname remis à zéro sauf le mots-clés
+
+//  Si y'a pas de mots-clés et une seule catégorie, est-ce qu'on affiche CATEGORIES: ECOUTER ?
+// Je pense que c’est bien de la faire, oui (tant qu’on peut donner une info sans nuire au parcours, on donne).
